Add example-contact-us package
A package that utilizes the VulcanEmail .send function to send a confirmation email when a guest wants contact with the 'site admins'. Useful for websites where the customer wants to be in contact with the 'site admins' (e.g. the HR or support team) 

The mutation check on collection.js simply returns the document. This might be too unsafe.

Permissions er perhaps too loose, I can't grasp what happens when not declared.