Fix broken favicon link
The feed was claiming a broken link as its favicon, causing Feedly to
show a big ugly letter L as a substitute/placeholder. Change it to point
at the Cloudinary-hosted favicon.

This is currently very-not-ideal because the icon is 16x16, and there
are a few contexts where Feedly wamts to display it at a much larger size.