Update permissions : movies.create -> movie.create
This was breaking the use of default mutations.
I also updated the name of the mutator used in seed.js: newMutation -> createMutator. I can see that there is an `action: 'movie.create` arg passed to the mutator, however it seems that it is not used in the mutator, so it seems useless from my point of view. Ok to remove @SachaG ?